requirements.txt
streamlit
requests
googlesearch-python
sherlock
holehe
theHarvester
spiderfoot
